Cleaned Up Code
Due to my code being such a mess, I completely changed up the HTML so that the CSS would be cleaner and I would better be able to make the design fully responsive. After cleaning up the HTML I completely did the CSS from the ground up. Because of having less classes in my HTML the CSS looks a lot cleaner and easier to read. I also opted for a more simple CSS reset to better fit my code.